refactor: split focus export row formatting
